Added method to config that will convert a Shotgun API key to a port number for dynamic host/port handling.
Changed the server and wrapper classes to generate host and port
dynamically. Uses the local host by default and generates the port using
the Shotgun API key.

Changed the DoubleBarrelServer to accept the same arguments as a
regular DoubleBarrel object upon initialization.

Separated the connecting to a separate function in the wrapper.